+++ src/components/Layout/Footer/Footer.module.scss
@@ -1,0 +1,68 @@
+@use 'src/styles/constants/colors';
+
+.footerWrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  padding: 40px 16px;
+  color: colors.$FOOTERFONTCOLOR;
+  background-color: colors.$FOOTERBGCOLOR;
+}
+
+.contentsWrapper {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  width: 100%;
+  max-width: 1040px;
+}
+
+.mainContents {
+  display: flex;
+  flex-direction: column;
+  margin-bottom: 0;
+}
+
+.subContents {
+  display: flex;
+  flex-direction: column;
+  margin-top: 10px;
+  font-size: 14px;
+  line-height: 20px;
+  text-align: right;
+
+  .subTitle {
+    font-size: 15px;
+  }
+
+  .subText {
+    margin-bottom: 5px;
+  }
+
+  .notification {
+    font-size: 20px;
+    font-weight: 700;
+  }
+}
+
+.mainText {
+  margin-top: 5px;
+  font-size: 13px;
+  line-height: 15px;
+}
+
+.title {
+  margin-bottom: 20px;
+  font-size: 17px;
+}
+
+@media screen and (max-width: 600px) {
+  .contentsWrapper {
+    flex-direction: column;
+  }
+
+  .subContents {
+    text-align: left;
+  }
+}
