--- src/components/tmpSearchForm/searchForm.module.scss
@@ -1,5 +1,0 @@
-.searchFormWrapper {
-  width: 300px;
-  height: 100px;
-  background-color: #ffffff;
-}
