+++ src/hooks/useAppDispatch.ts
@@ -1,5 +1,4 @@
 import { useDispatch } from 'react-redux'
 import type { AppDispatch } from 'states/index'
 
-// instead of plain `useDispatch`
 export const useAppDispatch = (): AppDispatch => useDispatch<AppDispatch>()
