--- src/routes/Search/Search.module.scss
@@ -1,50 +1,0 @@
-@use 'src/styles/constants/colors';
-
-.container {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-
-  .searchFormWrapper {
-    position: relative;
-    display: flex;
-    width: 90%;
-    max-width: 700px;
-    margin: 20px auto;
-    background-color: colors.$WHITE;
-    border-radius: 42px;
-
-    form {
-      display: flex;
-      width: 100%;
-    }
-
-    .inputWrapper {
-      display: flex;
-      flex: 1;
-      align-items: center;
-      padding: 20px 24px;
-      font-size: 18px;
-      font-weight: 400;
-      line-height: 1.6;
-
-      svg {
-        margin-right: 12px;
-      }
-
-      input {
-        width: 100%;
-      }
-    }
-
-    .searchButton {
-      padding: 18px 32px;
-      font-size: 18px;
-      font-weight: 700;
-      color: colors.$WHITE;
-      letter-spacing: -0.018em;
-      background-color: #007be9;
-      border-radius: 0 42px 42px 0;
-    }
-  }
-}
