+++ src/routes/SearchPage/Search/Dropdown/Dropdown.module.scss
@@ -26,11 +26,11 @@
   }
 
   .item {
-    overflow: hidden;
     display: flex;
     align-items: center;
     width: 100%;
     padding: 8px;
+    overflow: hidden;
     font-size: 16px;
     text-overflow: ellipsis;
     white-space: nowrap;
@@ -55,8 +55,9 @@
       fill: colors.$GRAYA;
     }
 
-    strong {
+    mark {
       font-weight: 900;
+      background-color: colors.$WHITE;
     }
   }
 }
