+++ src/routes/SearchPage/Search/Search.module.scss
@@ -1,0 +1,17 @@
+@use 'src/styles/constants/colors';
+
+.container {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+
+  .searchFormWrapper {
+    position: relative;
+    display: flex;
+    width: 90%;
+    max-width: 700px;
+    margin: 20px auto;
+    background-color: colors.$WHITE;
+    border-radius: 42px;
+  }
+}
