+++ src/routes/index.tsx
@@ -1,5 +1,6 @@
-import Layout from 'components/layout'
 import { Route, Routes } from 'react-router-dom'
+
+import Layout from 'components/Layout'
 import SearchPage from './SearchPage'
 
 const App = () => {
